requirements.txt
#
# This file is autogenerated by pip-compile with python 3.7
# To update, run:
#
#    pip-compile
#
aioredis==1.3.1
    # via channels-redis
appnope==0.1.2
    # via ipython
asgiref==3.4.1
    # via
    #   channels
    #   channels-redis
    #   daphne
    #   django
async-timeout==4.0.1
    # via aioredis
attrs==21.2.0
    # via
    #   automat
    #   service-identity
    #   twisted
autobahn==21.11.1
    # via daphne
automat==20.2.0
    # via twisted
backcall==0.2.0
    # via ipython
cassandra-driver==3.25.0
    # via django-cassandra-engine
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via cryptography
channels==3.0.4
    # via
    #   -r requirements.in
    #   channels-redis
channels-redis==3.3.1
    # via -r requirements.in
charset-normalizer==2.0.9
    # via requests
click==8.0.3
    # via
    #   geomet
    #   pip-tools
constantly==15.1.0
    # via twisted
cryptography==36.0.1
    # via
    #   autobahn
    #   pyopenssl
    #   service-identity
daphne==3.0.2
    # via channels
decorator==5.1.0
    # via ipython
django==3.2.10
    # via
    #   channels
    #   django-cassandra-engine
    #   django-extensions
    #   django-redis
django-cassandra-engine==1.6.3
    # via -r requirements.in
django-channels==0.7.0
    # via -r requirements.in
django-extensions==3.1.5
    # via -r requirements.in
django-redis==5.1.0
    # via -r requirements.in
geomet==0.2.1.post1
    # via cassandra-driver
hiredis==2.0.0
    # via aioredis
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
idna==3.3
    # via
    #   hyperlink
    #   requests
    #   twisted
importlib-metadata==4.9.0
    # via
    #   click
    #   pep517
incremental==21.3.0
    # via twisted
ipython==7.30.1
    # via -r requirements.in
jedi==0.18.1
    # via ipython
matplotlib-inline==0.1.3
    # via ipython
msgpack==1.0.3
    # via channels-redis
oauthlib==3.1.1
    # via requests-oauthlib
parso==0.8.3
    # via jedi
pep517==0.12.0
    # via pip-tools
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pip-tools==6.4.0
    # via -r requirements.in
prompt-toolkit==3.0.24
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.2.8
    # via service-identity
pycparser==2.21
    # via cffi
pygments==2.10.0
    # via ipython
pyopenssl==21.0.0
    # via twisted
pytz==2021.3
    # via django
redis==3.5.3
    # via django-redis
requests==2.26.0
    # via
    #   django-channels
    #   requests-oauthlib
requests-oauthlib==1.3.0
    # via django-channels
service-identity==21.1.0
    # via twisted
six==1.16.0
    # via
    #   automat
    #   cassandra-driver
    #   django-cassandra-engine
    #   django-channels
    #   geomet
    #   pyopenssl
    #   service-identity
sqlparse==0.4.2
    # via django
tomli==2.0.0
    # via pep517
traitlets==5.1.1
    # via
    #   ipython
    #   matplotlib-inline
twisted[tls]==21.7.0
    # via daphne
txaio==21.2.1
    # via autobahn
typing-extensions==4.0.1
    # via
    #   asgiref
    #   async-timeout
    #   importlib-metadata
    #   twisted
urllib3==1.26.7
    # via requests
wcwidth==0.2.5
    # via prompt-toolkit
wheel==0.37.0
    # via pip-tools
zipp==3.6.0
    # via
    #   importlib-metadata
    #   pep517
zope.interface==5.4.0
    # via twisted

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
